+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/generation/builders/IdentifiersBuilder.java
@@ -155,7 +155,7 @@ public class IdentifiersBuilder extends AbstractCodeBuilder implements
         }
 
         // The ArrayStore implementation is responsible for making a defensive copy of the index if needed
-        return array.getElement(assembler, node.sourcePosition(), index);
+        return array.getElement(assembler, node.sourcePosition(), index, false);
     }
 
     private ValueStore storeSubarrayAccess(AstSubarray node, ArrayStore array) {