+++ compiler/src/main/java/info/teksol/mc/mindcode/logic/arguments/arrays/ExternalArrayConstructor.java
@@ -39,7 +39,7 @@ public class ExternalArrayConstructor extends TablelessArrayConstructor {
     @Override
     public void expandInstruction(Consumer<LogicInstruction> consumer, Map<String, JumpTable> jumpTables) {
         AstContext astContext = instruction.getAstContext();
-        generateBoundsCheck(astContext, consumer, instruction.getIndex(), 1);
+        generateBoundsCheck(astContext, consumer, instruction.getIndex());
         switch (instruction) {
             case ReadArrInstruction rix -> consumer.accept(processor.createRead(astContext, rix.getResult(), memory, rix.getIndex()).setNonNegativeInt(rix.getIndex()));
             case WriteArrInstruction wix -> consumer.accept(processor.createWrite(astContext, wix.getValue(), memory, wix.getIndex()).setNonNegativeInt(wix.getIndex()));